backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
opencv-python==4.8.1.78
google-generativeai==0.3.2
Pillow==10.1.0
numpy==1.24.3
pymysql==1.1.0
sqlalchemy==2.0.23
pydantic-settings==2.1.0
pydantic[email]==2.5.0 